Print ArrayList successfully
Print an array list on one line with commas separating each value. Currently still includes the original text in the xml file.
Have two methods of printing. Commented out version on lines 38-39 prints each element on one line. Used an alternate one, as mentioned above.

Tried to move printing functionality to DisplayData class, but seems like you may not be able to access the Views of another file.
Instead, may need to be a "Fragment" of an Activity in order to access its views.

Implementing DisplayData as a Fragment is marked as a TODO item.