Added Bluetooth Activity and startup checks
Created a Bluetooth activity that links to the main activity as its parent.

Created some checks to see if Bluetooth is enabled and if not, to enable it.

Need to finalize the check that searches if Bluetooth device has been paired in the past, so that re-pairing doesn't have to occur every time.